pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitlab-to-github-migrator"
version = "0.1.0"
description = "Python tool for migrating GitLab projects to GitHub with full metadata preservation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyGithub==2.7.0",
    "python-gitlab==6.2.0",
    "requests==2.32.4",
]
# https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    # Include this to avoid accidentally publishing to PyPI:
    "Private :: Do Not Upload",
]

[project.scripts]
gitlab-to-github-migrator = "gitlab_to_github_migrator:main"
cleanup_test_repos = "gitlab_to_github_migrator.cleanup_test_repos:main"

[tool.uv]
dev-dependencies = [
    "basedpyright>=1.31.2",
    "codespell>=2.4.1",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.6.1",
    "ruff>=0.12.9",
]

[tool.ruff]
# Set as desired, typically 88 (black standard) or 100 (wide).
line-length = 119

[tool.ruff.lint]
select = [
    # See: https://docs.astral.sh/ruff/rules/
    "A",
    "AIR",
    "ANN",
    "ARG",
    "ASYNC",
    "B", # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "BLE",
    "C4",
    "COM", # https://docs.astral.sh/ruff/rules/#flake8-commas-com
    # "D", # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "DTZ",
    "E", # https://docs.astral.sh/ruff/rules/#error-e
    "EM",
    "EXE",
    "F", # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "FA",
    "FAST",
    "FBT",
    "FLY",
    "FURB",
    "G",
    "I", # https://docs.astral.sh/ruff/rules/#isort-i
    "ICN",
    "INP",
    "ISC",
    "LOG",
    "N",
    "NPY",
    "PD",
    "PERF",
    "PGH",
    "PIE",
    "PL",
    "PTH",
    "Q", # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM", # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
    "SLF",
    "SLOT",
    "T20",
    "TC",
    "TID",
    "TRY",
    "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
]
ignore = [
    # Disable some rules that are overly pedantic. Add/remove as desired:
    "ANN001", # Missing type annotation for function argument - too verbose for tests
    "BLE001", # Do not catch blind exception - sometimes necessary for robustness
    "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
    "COM819", # https://docs.astral.sh/ruff/rules/prohibited-trailing-comma/
    "D206", # https://docs.astral.sh/ruff/rules/docstring-tab-indentation/
    "D300", # https://docs.astral.sh/ruff/rules/triple-single-quotes/
    "E111", # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple/
    "E114", # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple-comment/
    "E117", # https://docs.astral.sh/ruff/rules/over-indented/
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E501", # https://docs.astral.sh/ruff/rules/line-too-long/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    "G004",   # Logging statement uses f-string - f-strings are fine for logging
    "ISC002", # https://docs.astral.sh/ruff/rules/multi-line-implicit-string-concatenation/
    "PD008",  # Allow the use of .at[]
    "PD009",  # Allow the use of .iat[]
    "PD901",  # Allow the use of df as var name
    "PLC0415", # https://docs.astral.sh/ruff/rules/import-outside-top-level/
    "PLR0913", # Too many arguments in function definition - sometimes necessary
    "PLR2004", # Magic value used in comparison - constants not always needed
    "Q000", # https://docs.astral.sh/ruff/rules/bad-quotes-inline-string/
    "Q001", # https://docs.astral.sh/ruff/rules/bad-quotes-multiline-string/
    "Q002", # https://docs.astral.sh/ruff/rules/bad-quotes-docstring/
    "Q003", # https://docs.astral.sh/ruff/rules/avoidable-escaped-quote/
    "S101",  # Allow the use of assert
    # "S110",  # try-except-pass detected - sometimes appropriate
    "S311",  # Standard pseudo-random generators not suitable for crypto - using for test names only
    "S607",  # Starting a process with a partial executable path - git/subprocess are safe
    "SIM105", # Use contextlib.suppress - try-except-pass is clearer sometimes
    "SIM108",  # Allow if-else-block instead of ternary if-exp
    "TC006", # Allow non-quoted types in type cast https://docs.astral.sh/ruff/rules/runtime-cast-value/
    "TRY003", # Avoid specifying long messages outside the exception class
    "TRY301", # Abstract raise to an inner function - not needed for simple cases
    "W191", # https://docs.astral.sh/ruff/rules/tab-indentation/
]

[tool.basedpyright]
# BasedPyright currently seems like the best type checker option, much faster
# than mypy and with a good extension for VSCode/Cursor.
# https://marketplace.visualstudio.com/items?itemName=detachhead.basedpyright
# https://docs.basedpyright.com/latest/configuration/config-files/#sample-pyprojecttoml-file
include = ["src", "tests"]

allowedUntypedLibraries = ["gitlab"]

# By default BasedPyright is very strict, so you almost certainly want to disable
# some of the rules.
# First, these turn off warnings about (yes) how you ignore warnings:
reportIgnoreCommentWithoutRule = false
reportUnnecessaryTypeIgnoreComment = false
# A few typically noisy warnings are next.
# How many you enable is up to you. The first few are off by default, but you can
# comment/uncomment these as desired:
reportMissingTypeStubs = false
reportUnusedCallResult = false
reportAny = false
reportExplicitAny = false
reportImplicitStringConcatenation = false
# reportUnreachable = false
reportPrivateImportUsage = false
reportPrivateLocalImportUsage = false
reportUnusedImport = false  # Done by Ruff F401
# reportMissingImports = false
# reportUnnecessaryIsInstance = false
# reportUnknownVariableType = false
# reportUnknownArgumentType = false
reportUnknownMemberType = false  # gitlab has no type hints
reportUnknownLambdaType = false  # gitlab has no type hints
# reportMissingParameterType = false
# reportUnknownParameterType = false
# reportUnannotatedClassAttribute = false
# reportOptionalMemberAccess = false
# reportPossiblyUnboundVariable = false
# reportUninitializedInstanceVariable = false

[tool.hatch.build.targets.wheel]
packages = ["src/gitlab_to_github_migrator"]

[tool.codespell]
# Add here as needed:
# ignore-words-list = "foo,bar"
# skip = "foo.py,bar.py"

[tool.pytest.ini_options]
python_files = ["*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = [
    "tests",
]
pythonpath = ["src"]
norecursedirs = []
filterwarnings = []
